# backend/swagger.yaml
openapi: 3.0.0
paths:
  /oauth/refresh:
    get:
      operationId: OauthController_getNewToken
      summary: Get new refresh and access token
      description: get new access token and refressh token to can continue to do request
      parameters: []
      tags:
        - Oauth
      responses:
        '200':
          description: 'The refresh work, new accessToken is return and new refreshToken is  set in httpOnly with /oauth path'
          content:
            application/json:
              shema:
                type: object
                required:
                  - accessToken
                properties:
                  accessToken:
                    type: string
        '403':
          description: '403 is returned when: <br/>- ''refreshToken'' cookie is not set <br/> -  It''s false refresh token <br/>- The refresh token is expire <br/> -  The refreshToken is already use (disconectFromEvryWhere is launch)'
  /oauth/singin:
    post:
      operationId: OauthController_signup
      summary: Connect user to the website
      description: Use 42 code in the body to create account or connect the user if is already exist
      parameters:
        - in: body
          name: code
          description: The code return by oauth 42
          schema:
            type: object
            required:
              - code
            properties:
              code:
                type: string
      tags:
        - Oauth
      responses:
        '201':
          description: Add refreshToken in cokkie in HTTPOnly in path /oauth and give access token and return if is you're first connection
          content:
            application/json:
              schema:
                type: object
                required:
                  - accessToken
                  - newUser
                properties:
                  accessToken:
                    type: string
                  newUser:
                    type: boolean
        '400':
          description: 400 is return if the code is undefined or if is bad or if the 42 api failed
...
  /oauth/disconect:
    delete:
      operationId: OauthController_disconect
      description: disable the refresh token in the cookie
      summary: Disconect user
      parameters: []
      tags:
        - Oauth
      responses:
        '200':
          description: The disconection work
        '403':
          description: '403 is returned when: <br/>- ''refreshToken'' cookie is not set <br/> -  It''s false refresh token <br/>- The refresh token is expire <br/> -  The refreshToken is already use (disconectFromEvryWhere is launch)'
  /oauth/disconectEveryWhere:
    delete:
      operationId: OauthController_disconectEveryWhere
      summary: Disconect user from every navigator
      description: disable all refresh token of the user
      parameters: []
      tags:
        - Oauth
      responses:
        '200':
          description: The disconection work
        '403':
          description: '403 is returned when: <br/>- ''refreshToken'' cookie is not set <br/> -  It''s false refresh token <br/>- The refresh token is expire <br/> -  The refreshToken is already use (disconectFromEvryWhere is launch)'
...
info:
  title: TRANSCENDENCE
  description: Our 42 project ft_transcendence API description.
  version: v1
  contact: {}
...
servers: []
components:
  securitySchemes:
    TransandanceToken:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Enter access token for transandance
...